Refactor: move SignException
